app/rest_api/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
django-cors-headers>=3.5
djangorestframework
markdown
django-filter
cryptography
django-rest-knox>=4.0.1
